Fix "No such file or directory" error for Powerline
Previous configuration was set to source powerline.sh. This resulted in
the following error:

-bash: ./Library/Python/3.7/lib/python/site-packages/powerline/bindings/bash/../../../scripts/powerline-config: No such file or directory

Not only did it prevent powerline-status from launching BUT it prevented
the configuration from loading if launched manually. Searched and found
the problem referenced in "https://github.com/powerline/powerline/issues/1960".
Chose not to symlink and determined this was not a $PATH issue.
Followed link to
"https://askubuntu.com/questions/283908/how-can-i-install-and-use-powerline-plugin/953875#953875"
and used snippet to fix the problem.